requirements.txt
torch>=1.13.0
numpy>=1.21.0
Pillow>=8.0.0
typing-extensions>=4.0.0
requests>=2.25.0